refactor: 위치 리팩터링
- 매직넘버 상수화
- 메서드 추출
- 좌표값 형식 검증
- 형식 검증 테스트 코드 추
